+++ src/Util/ConfigurationInterface.php
@@ -14,6 +14,12 @@ namespace League\CommonMark\Util;
 interface ConfigurationInterface
 {
     /**
+     * @internal
+     * @deprecated
+     */
+    public const MISSING = '833f2700-af8d-49d4-9171-4b5f12d3bfbc';
+
+    /**
      * Merge an existing array into the current configuration
      *
      * @param array<string, mixed> $config