+++ tests/unit/Block/Parser/ListParserTest.php
@@ -118,7 +118,10 @@ final class ListParserTest extends TestCase
         $this->assertSame('+', $container->getListData()->bulletChar);
     }
 
-    public function testUnorderedListWithCustomMarker()
+    /**
+     * @dataProvider configPaths
+     */
+    public function testUnorderedListWithCustomMarker(string $configPath)
     {
         $input = '^ Foo';
 
@@ -126,8 +129,11 @@ final class ListParserTest extends TestCase
         $context->setNextLine($input);
         $cursor = new Cursor($input);
 
+        $config = new Configuration();
+        $config->set($configPath, ['^']);
+
         $parser = new ListParser();
-        $parser->setConfiguration(new Configuration(['unordered_list_markers' => ['^']]));
+        $parser->setConfiguration($config);
         $this->assertTrue($parser->parse($context, $cursor));
 
         $container = $context->getContainer();
@@ -138,7 +144,10 @@ final class ListParserTest extends TestCase
         $this->assertSame('^', $container->getListData()->bulletChar);
     }
 
-    public function testUnorderedListWithDisabledMarker()
+    /**
+     * @dataProvider configPaths
+     */
+    public function testUnorderedListWithDisabledMarker(string $configPath)
     {
         $input = '+ Foo';
 
@@ -146,12 +155,18 @@ final class ListParserTest extends TestCase
         $context->setNextLine($input);
         $cursor = new Cursor($input);
 
+        $config = new Configuration();
+        $config->set($configPath, ['-', '*']);
+
         $parser = new ListParser();
-        $parser->setConfiguration(new Configuration(['unordered_list_markers' => ['-', '*']]));
+        $parser->setConfiguration($config);
         $this->assertFalse($parser->parse($context, $cursor));
     }
 
-    public function testInvalidListMarkerConfiguration()
+    /**
+     * @dataProvider configPaths
+     */
+    public function testInvalidListMarkerConfiguration(string $configPath)
     {
         $this->expectException(\RuntimeException::class);
         $this->expectExceptionMessage('Invalid configuration option "unordered_list_markers": value must be an array of strings');
@@ -161,9 +176,21 @@ final class ListParserTest extends TestCase
         $context->setNextLine($input);
         $cursor = new Cursor($input);
 
+        $config = new Configuration();
+        $config->set($configPath, '-');
+
         $parser = new ListParser();
-        $parser->setConfiguration(new Configuration(['unordered_list_markers' => '-']));
+        $parser->setConfiguration($config);
 
         $parser->parse($context, $cursor);
+    }
+
+    /**
+     * @return iterable<string[]>
+     */
+    public function configPaths(): iterable
+    {
+        yield ['unordered_list_markers'];
+        yield ['commonmark/unordered_list_markers'];
     }
 }